requirements.txt
Django>=4.1.1,<4.2
djangorestframework>=3.14.0,<3.15
psycopg2>=2.8.6,<3.0
drf-spectacular>=0.26.0,<0.27
django-cors-headers>=4.2,<4.3
